services/raw_data_service/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
psycopg2-binary
geoalchemy2
httpx
gtfs-kit
protobuf
pandas
python-dotenv
apscheduler
